Adding module that collects system information. There are several sources:   1. The 'inxi' tool (must be on a 'path'): inxi -Fbfrlp;   2. Content of a '/proc/cpuinfo' file;   3. Content of a '/proc/meminfo' file;   4. Output of a 'lscpu' command;   5. Output of a 'nvidia-smi -q' command;
For now, the goal is to collect all info that may be useful in the future. The results are serialized as JSON objects in 'hw' namespace in log files: hw.inxi, hw.cpuinfo, hw.meminfo, hw.lscpu and hw.nvidiasmi.
This functionality is disabled by default. To enable it, use exp.sys_info command line argument. It's a comma separated string of tools' identifiers that user is intrested in, for instance:
-Pexp.sys_info='"inxi,lscpu,nvidiasmi"' to collect output of inxi, lscpu and nvidia-smi tools.