Fixing validator docker image check.
There's a sequence of parameters that defines how docker image is computed.
In validator.py list of benchmarks is processed with Processor, and this
means all parameters have been computed. This means that the actual parameter
that contains docker image to use is exp.docker_image. BTW other, framework
specific parameters, such as nvcnn.docker_image may contain same or different
values, but they are not used.

So, instead of this code:
   docker_img_key = '%s.docker_image' % (exp['exp.framework'])
the project now uses this:
   docker_img_key = 'exp.docker_image'