Latest TensorFlow benchmark backen 'nvtfcnn'.
This TensorFlow benchmark backend is based on latest script examples in
NGC containers. It provides state of the art performance for many models
using 8 GPUs in one box.

This update is intended to introduce this benchmark backend that should
be used by default for benchmarking hardware. Later in 2018, we will
update nvtfcnn benchmark backend introducing multi-node functionality.

Consider current implementation as a temporary (introductory) nvtfcnn
integration with DLBS.

 - Backend name is "nvtfcnn": -Pexp.framework='"nvtfcnn"'
 - Docker image is nvcr.io/nvidia/tensorflow:18.07-py3
 - Fully connected models are not supported.
 - Only docker runtime is supported.