Fixing TensorRT docker files.
On some systems with docker using overlay storage driver, it is not possible to
remove folder (rm -rf DIR) created in other layer.
If you see messages like this:
   rm: cannot remove DIR: Directory not empty.
Run the following:
   docker info
Most likely, you'll see that your storage driver is `overlay`.

This update fixes this error by not removing a root folder of the project. Instead,
it removes a build folder which is created in the same layer.

More info on the problem is here: https://github.com/moby/moby/issues/27214