Fixing PYTHONPATH issue during post-processing phase.
In certain benchmarks that are based on external benchmark backend, a post-processing is required to ensure that all log files contain required information such as achieved throughput.
The python module responsible for this is called from launcher scripts where PYTHONPATH is not properly set what resulted in "module cannot be loaded" error. It's fixed now temporarily by adjusting sys.path on the fly.